Refactor bitches

+++ src/mpi/cmatrix.c
@@ -5,7 +5,7 @@
 int mpi_send_items(
     Item const* const items, int len, int dest, int tag, MPI_Comm comm) {
     unsigned char* buf = (unsigned char*) items;
-    int buf_len = len * sizeof(Item);
+    int buf_len = len * (int) sizeof(Item);
     return MPI_Send(buf, buf_len, MPI_UNSIGNED_CHAR, dest, tag, comm);
 }
 
@@ -17,6 +17,6 @@ int mpi_recv_items(
     MPI_Comm comm,
     MPI_Status* status) {
     unsigned char* buf = (unsigned char*) items;
-    int buf_len = len * sizeof(Item);
+    int buf_len = len * (int) sizeof(Item);
     return MPI_Recv(buf, buf_len, MPI_UNSIGNED_CHAR, source, tag, comm, status);
 }
